dispatch event when item is opened/closed for editing; add client api method to retrieve content profile by id
