+++ scripts/apps/authoring/widgets/widgets.ts
@@ -1,9 +1,9 @@
 import {debounce} from 'lodash';
-import {IContentProfile} from 'superdesk-interfaces/ContentProfile';
 import {isWidgetVisibleForContentProfile} from 'apps/workspace/content/components/WidgetsConfig';
 import {gettext} from 'core/utils';
 import {isKilled} from 'apps/archive/utils';
 import {AuthoringWorkspaceService} from '../authoring/services/AuthoringWorkspaceService';
+import {IContentProfile} from 'superdesk-api';
 
 function AuthoringWidgetsProvider() {
     var widgets = [];
